move schemas/* to schemas.ts
